requirements.txt
flask==1.1.4
flask-restx
pytest
pyyaml
